Changed behavior: aligned option is now required if you want the initial alignment to be used for the initial tree setting.
Previously, it was ignoring that setting and just using whether or not all of
the sequences have the same length as the trigger to detect when it should
use the initial alignment.

This commit also handles that '--aligned' command line setting in the GUI, so
that the user can choose to use or ignore the initial alignment.