Modifications to aligner and merger tools in tool.py, and Alignment in alignment.py. The changes fix multiple bugs in how multi-locus data was being handled.
1) When some loci's fasta files were missing taxa, decomps could
result in empty alignments, which were not handled by the aligner
tools, which led to errors. 2) When some loci's fasta files contained
taxa with no data (i.e. all "-"s), the aligner tools would create
input alignment files with empty sequences. 3) When decomp resulted
in singleton alignments, the gaps were not removed before the sequence
was given to the FakeJob as results. This led to odd alignments that
had columns that were all gaps, which often led to very odd branch
length estimates.

With the addition of an Alignment.unalign method and reworking of
the aligner tools' create_job and _prepare_input methods, these
problems should now be fixed.