Created 'tree-estimator-model' commandline option.
Created commandline option to allow user to specify the model
for the tree estimator tool as a string. I did this directly in
mainsate.sate_main(), because I did not want to make this option
part of sate configuration. It is a hack to allow the model to
be passed at the commandline. In general, this should always be
done via the config file.