Attempting fix of recursion directory bug.
During a tree decomposition, if both subtrees have multiple
recursion steps, the 'recursion_index' of one of the SateAlignerJob
instances falls behind. This causes it to try to create the same
temporary recursion directory a second time, which triggers an
OSError.

This commit makes the 'recursion_index' a global variable of the
SateAlignerJob class, which should fix this bug.