Updating starting tree parsing in mainsate.
I am adding a 'starting_tree=True' flag to the
read_and_encode_splits call within the finish_sate_execution function.
This will cause the tip labels to be checked against the sequences
and will result in an TaxaLabelsMismatchError if there is a
mismatch.

Adding associated test to test_main.py.