Updating get_invoke_run_sate_command.
sate.configure.get_invoke_run_sate_command() is used by
run_sate_gui.py to invoke the CLI script. When the platform was
not Windows or Darwin, this returned "run_sate.py" as the path
to the CLI script. Thus, if you tried to run the GUI on a Linux
machine and were not in the same directory as the CLI script, the
GUI choked. I modifed this to return to full path to the CLI
script.

This function now returns the same info if the platform is Darwin
or "other" (i.e., not Windows or Darwin), so the logic can probably
be simplified.