Updating sate GUI script (run_sate_gui.py)
Adding an '_encode_arg' method to SateFrame class. This method
checks to see if the arg is a unicode instance, and if so, converts
it to a string using an encoding ('utf-8' by default). This method
is now called when handling all the GUI text fields.

This change should prevent the GUI from choking when unicode
characters are included in paths.