Adding conversion of RNA inputs to DNA.
Updating sate.mainsate.finish_sate_execution() function to convert all
RNA sequence data to DNA characters. The motivation is to make sure all
nucleotide data within a SATe analysis are encoded as DNA. This will
minimize the chance of tools choking on, or doing something unexpected
with, RNA data.  The RNA characters are restored before the final
alignments are written. These conversions are done with the
convert_rna_to_dna and convert_dna_to_rna methods of the
MultiLocusDataset class. The conversions are performed at the same
points when the taxon names are converted.

Currently, the config.commandline.datatype remains as RNA, which is used
to check if conversion back to DNA is necessary before the final outputs
are written. Thus, the data type can be 'RNA' internally, but all
nucleotide characters should be DNA.