Adding tests to test_main.py.
Adding a lot of tests, most of which check behaviors for RNA data.
Importantly, one tests whether the alignments are the same for
identical RNA and DNA sequence datasets. Also adding test to
ensure mixed data cannot be run currently.

Also overhauling a lot of the test code in SateTestCase and
test_alignment.py.