Updating GUI to override max subproblem defaults.
There are two controls for the max subproblem for the tree
decomposition, frac and size, both of which have defaults, and the
larger of the two is used to determine the maximum subset size
during decomp. This caused confusion from the GUI, because the
selected option would be ignored if it was smaller than the default
for the alternative option. The GUI now explicitly configures BOTH
options so that what the user sees is what they get.